+++ CHANGELOG.md
@@ -144,6 +144,7 @@
 - PR #4199 Fix potential race condition in memcpy_block
 - PR #4215 Fix performance regression in strings::detail::concatenate
 - PR #4214 Alter ValueError exception for GPU accelerated Parquet writer to properly report `categorical` columns are not supported.
+- PR #4232 Fix handling empty tuples of children in string columns
 
 
 # cuDF 0.12.0 (04 Feb 2020)